+++ Server/src/main/java/ForMZ/Server/domain/post/Entity/Post.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.post.Entity;
+
+public class Post {
+}
