--- Server/src/main/java/ForMZ/Server/domain/post/api/PostApi.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.post.api;
-
-public class PostApi {
-}
