--- Server/src/main/java/ForMZ/Server/domain/post/application/PostService.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.post.application;
-
-public class PostService {
-}
