+++ Server/src/main/java/ForMZ/Server/domain/post/controller/PostController.java
@@ -1,0 +1,9 @@
+package ForMZ.Server.domain.post.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class PostController {
+}
