--- Server/src/main/java/ForMZ/Server/domain/post/dao/PostRepository.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.post.dao;
-
-public class PostRepository {
-}
