--- Server/src/main/java/ForMZ/Server/domain/post/domain/Post.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.post.domain;
-
-public class Post {
-}
