+++ Server/src/main/java/ForMZ/Server/domain/post/repository/PostRepository.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.post.repository;
+
+public class PostRepository {
+}
