+++ Server/src/main/java/ForMZ/Server/domain/post/service/PostService.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.post.service;
+
+public class PostService {
+}
