--- Server/src/main/java/ForMZ/Server/domain/user/api/UserApi.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.user.api;
-
-public class UserApi {
-}
