--- Server/src/main/java/ForMZ/Server/domain/user/application/UserService.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.user.application;
-
-public class UserService {
-}
