+++ Server/src/main/java/ForMZ/Server/domain/user/controller/UserController.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.user.controller;
+
+public class UserController {
+}
