--- Server/src/main/java/ForMZ/Server/domain/user/dao/UserRepository.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.user.dao;
-
-public class UserRepository {
-}
