--- Server/src/main/java/ForMZ/Server/domain/user/domain/User.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.user.domain;
-
-public class User {
-}
