+++ Server/src/main/java/ForMZ/Server/domain/user/dto/UserUpdateReq.java
@@ -1,4 +1,4 @@
 package ForMZ.Server.domain.user.dto;
 
-public class UserProfileUpdateReq {
+public class UserUpdateReq {
 }
