+++ Server/src/main/java/ForMZ/Server/domain/user/entity/User.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.user.entity;
+
+public class User {
+}
