+++ Server/src/main/java/ForMZ/Server/domain/user/repository/UserRepository.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.user.repository;
+
+public class UserRepository {
+}
