+++ Server/src/main/java/ForMZ/Server/domain/user/service/UserService.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.user.service;
+
+public class UserService {
+}
